Reorganised project structure and added necessary files
