+++ .gitignore
@@ -1,1 +1,5 @@
 .DS_Store
+
+/target/
+*.class
+*.jar
