+++ src/main/java/io/github/Shiwang-Jais/mlutil/matrix/Matrix.java
@@ -1,4 +1,4 @@
-package players.genetic;
+package io.github.shiwangjais.mlutil.matrix;
 
 import java.util.ArrayList;
 import java.util.List;
